[MovieFileSearcher] Run searcher in another thread
This change reworks the file searcher (again).  It now runs in another
thread, or rather: The jobs for each directory run in another thread.
We had to make the database access work with multiple threads and fix
some other multithreading issues.  But this was worth the cost because
now MediaElch no longer freezes for reloading movies from slow disks.

It also makes it possible to remove calls to
`QApplication::processEvents()` that previously were the causes for
crashes.